refactor: change project folders structure
