+++ src/components/authorization/forms/signup-form.ts
@@ -1,18 +1,18 @@
-import BaseComponent from '../base-component/base-component';
-import Button from '../base-component/button/button';
-import NavigationLink from '../base-component/link/link';
-import Input from '../base-component/text-input-and-label/text-input';
+import BaseComponent from '../../base-component/base-component';
+import Button from '../../base-component/button/button';
+import NavigationLink from '../../base-component/link/link';
+import Input from '../../base-component/text-input-and-label/text-input';
 import './form.css';
-import Routes from '../../app/router/router.types';
-import { Errors, SignUp, Token } from '../../app/loader/loader.types';
-import { createUser } from '../../app/loader/services/user-services';
-import { setDataToLocalStorage } from '../../utils/local-storage/local-storage';
+import Routes from '../../../app/router/router.types';
+import { Errors, SignUp, Token } from '../../../app/loader/loader.types';
+import { createUser } from '../../../app/loader/services/user-services';
+import { setDataToLocalStorage } from '../../../utils/local-storage/local-storage';
 import { GoogleBtnClasses, GoogleBtnTypes } from '../google-button/google-btn.types';
 import GoogleButton from '../google-button/google-btn';
-import { VALID_EMAIL, VALID_NAME, VALID_PASSWORD } from '../../utils/consts';
+import { VALID_EMAIL, VALID_NAME, VALID_PASSWORD } from '../../../utils/consts';
 import { InputConflictMessages, ValidityMessages } from './form.types';
-import Select from '../select/select';
-import { convertRegexToPattern } from '../../utils/utils';
+import Select from '../../base-component/select/select';
+import { convertRegexToPattern } from '../../../utils/utils';
 
 export default class SignupForm extends BaseComponent<'form'> {
   private formHeader: BaseComponent<'h4'> = new BaseComponent(
