+++ src/components/authorization/google-button/google-btn.types.ts
@@ -1,4 +1,4 @@
-import { LogIn, SignUp } from '../../app/loader/loader.types';
+import { LogIn, SignUp } from '../../../app/loader/loader.types';
 
 export enum GoogleBtnClasses {
   Signup = 'g_id_signup',
