+++ src/map/google.visualization/google.visualization.ts
